+++ pkgs/test/test/runner/compiler_runtime_matrix_test.dart
@@ -22,8 +22,9 @@ void main() {
   for (var runtime in Runtime.builtIn) {
     for (var compiler in runtime.supportedCompilers) {
       // Ignore the platforms we can't run on this OS.
-      if (runtime == Runtime.internetExplorer && !Platform.isWindows ||
-          runtime == Runtime.safari && !Platform.isMacOS) {
+      if (((runtime == Runtime.internetExplorer || runtime == Runtime.edge) &&
+              !Platform.isWindows) ||
+          (runtime == Runtime.safari && !Platform.isMacOS)) {
         continue;
       }
       group('--runtime ${runtime.identifier} --compiler ${compiler.identifier}',